Fixed addnode

+++ Estroimpl.h
@@ -9,5 +9,5 @@ namespace Estro {
 	void init(unsigned int width, unsigned int height, std::string title);
 
 	bool isLooping();
-	void drawRectangle(int x, int y, int w, int h);
+	void drawRectangle(float x, float y, float w, float h);
 }
